fix 金币使用int计算，解决浮点精度问题

+++ model/order.go
@@ -1,22 +1,22 @@
 package model
 
 type Order struct {
-	Id            int64   `structs:"id" json:"orderId"`
-	UserId        int64   `structs:"user_id" json:"userId" validate:"number"`
-	ServiceId     int64   `structs:"service_id" json:"serviceId" validate:"required,number"`
-	AdvisorId     int64   `structs:"advisor_id" json:"advisorId" validate:"required,number"`
-	Situation     string  `structs:"situation" json:"situation" validate:"required,max=3000"`
-	Question      string  `structs:"question" json:"question" validate:"required,max=200"`
-	Coin          float32 `structs:"coin" json:"coin" validate:""`
-	RushCoin      float32 `structs:"rush_coin" json:"rushCoin"`
-	CreateTime    int64   `structs:"create_time" json:"createTime"`
-	RushTime      int64   `structs:"rush_time" json:"rushTime"`
-	Reply         string  `structs:"reply" json:"reply" validate:""`
-	Status        int     `structs:"status"`
-	Rate          int     `structs:"rate" json:"rate" validate:""`
-	Comment       string  `structs:"comment" json:"comment" validate:""`
-	CommentTime   int64   `structs:"comment_time" json:"commentTime"`
-	CommentStatus int64   `structs:"comment_status" json:"commentStatus"`
+	Id            int64  `structs:"id" json:"orderId"`
+	UserId        int64  `structs:"user_id" json:"userId" validate:"number"`
+	ServiceId     int64  `structs:"service_id" json:"serviceId" validate:"required,number"`
+	AdvisorId     int64  `structs:"advisor_id" json:"advisorId" validate:"required,number"`
+	Situation     string `structs:"situation" json:"situation" validate:"required,max=3000"`
+	Question      string `structs:"question" json:"question" validate:"required,max=200"`
+	Coin          int64  `structs:"coin" json:"coin" validate:""`
+	RushCoin      int64  `structs:"rush_coin" json:"rushCoin"`
+	CreateTime    int64  `structs:"create_time" json:"createTime"`
+	RushTime      int64  `structs:"rush_time" json:"rushTime"`
+	Reply         string `structs:"reply" json:"reply" validate:""`
+	Status        int    `structs:"status"`
+	Rate          int    `structs:"rate" json:"rate" validate:""`
+	Comment       string `structs:"comment" json:"comment" validate:""`
+	CommentTime   int64  `structs:"comment_time" json:"commentTime"`
+	CommentStatus int64  `structs:"comment_status" json:"commentStatus"`
 }
 type OrderInitInfo struct {
 	UserId    int64  `structs:"user_id" json:"userId" validate:"number"`
@@ -26,12 +26,12 @@ type OrderInitInfo struct {
 	Question  string `structs:"question" json:"question" validate:"required,max=200"`
 }
 type OrderReply struct {
-	Id        int64   `structs:"id" json:"orderId" validate:"min=1"`
-	AdvisorId int64   `structs:"advisor_id" json:"advisorId"`
-	Reply     string  `structs:"reply" json:"reply" validate:"min=1200,max=5000"`
-	Coin      float32 `structs:"coin" json:"coin" validate:""`
-	RushCoin  float32 `json:"rushCoin"`
-	Status    int64   `json:"status"`
+	Id        int64  `structs:"id" json:"orderId" validate:"min=1"`
+	AdvisorId int64  `structs:"advisor_id" json:"advisorId"`
+	Reply     string `structs:"reply" json:"reply" validate:"min=1200,max=5000"`
+	Coin      int64  `structs:"coin" json:"coin" validate:""`
+	RushCoin  int64  `json:"rushCoin"`
+	Status    int64  `json:"status"`
 }
 type OrderRush struct {
 	Id       int64 `struct:"id" json:"orderId"`
